lockdown delete item | frontend & backend

+++ backend/src/resolvers/Mutation.js
@@ -49,9 +49,19 @@ const Mutations = {
     async deleteItem(parent, args, ctx, info) {
         const where = { id: args.id };
         // 1.find item
-        const item = await ctx.db.query.item({ where }, `{id title}`);
+        const item = await ctx.db.query.item(
+            { where },
+            `{id title user { id }}`
+        );
         // 2.check if the own that item/have permissions
-        // TODO
+        const ownsItem = item.user.id === ctx.request.userId;
+        const hasPermissions = ctx.request.user.permissions.some((permission) =>
+            ['ADMIN', 'ITEMDELETE'].includes(permission)
+        );
+
+        if (!ownsItem || !hasPermissions) {
+            throw new Error("You don't have permission to do that");
+        }
         // 3.delete it!
         return ctx.db.mutation.deleteItem({ where }, info);
     },
